Move audio interfaces to own folder
