+++ OsuPlayer/Modules/Audio/Interfaces/IHasDiscordRpc.cs
@@ -1,0 +1,6 @@
+﻿namespace OsuPlayer.Modules.Audio.Interfaces;
+
+public interface IHasDiscordRpc
+{
+    public void DisposeDiscordClient();
+}
