+++ OsuPlayer/Modules/Audio/Interfaces/IPlayModes.cs
@@ -1,0 +1,9 @@
+﻿using OsuPlayer.Data.OsuPlayer.Enums;
+
+namespace OsuPlayer.Modules.Audio.Interfaces;
+
+public interface IPlayModes
+{
+    public Bindable<bool> IsShuffle { get; }
+    public Bindable<RepeatMode> RepeatMode { get; }
+}
