+++ OsuPlayer/Program.cs
@@ -4,6 +4,7 @@ using Avalonia.ReactiveUI;
 using OsuPlayer.Extensions;
 using OsuPlayer.IO.Importer;
 using OsuPlayer.Modules.Audio.Engine;
+using OsuPlayer.Modules.Audio.Interfaces;
 using OsuPlayer.Modules.Services;
 using OsuPlayer.Windows;
 using Splat;
