+++ OsuPlayer/Views/BlacklistEditorViewModel.cs
@@ -5,6 +5,7 @@ using DynamicData;
 using OsuPlayer.Base.ViewModels;
 using OsuPlayer.Data.OsuPlayer.StorageModels;
 using OsuPlayer.IO.Storage.Blacklist;
+using OsuPlayer.Modules.Audio.Interfaces;
 using ReactiveUI;
 
 namespace OsuPlayer.Views;
