+++ OsuPlayer/Views/SettingsViewModel.cs
@@ -6,6 +6,7 @@ using Avalonia.Media;
 using OsuPlayer.Base.ViewModels;
 using OsuPlayer.Data.OsuPlayer.Classes;
 using OsuPlayer.Data.OsuPlayer.Enums;
+using OsuPlayer.Modules.Audio.Interfaces;
 using OsuPlayer.Modules.Services;
 using OsuPlayer.Network;
 using OsuPlayer.Styles;
